+++ pkg/config/config.go
@@ -58,6 +58,23 @@ func ParseHelmInstallTimeoutAnnotation(raw string) (time.Duration, error) {
 	return d, nil
 }
 
+// ParsePositiveDuration parses raw as a time.Duration and rejects malformed
+// or non-positive values. Flux HelmRelease fields (Interval, Timeout,
+// RetryInterval) require strictly positive durations, so a misconfigured
+// flag must fail fast at startup rather than propagating into every HR.
+// Shared between cozystack-operator and cozystack-api so both paths reject
+// the same set of inputs at startup.
+func ParsePositiveDuration(flagName, raw string) (time.Duration, error) {
+	d, err := time.ParseDuration(raw)
+	if err != nil {
+		return 0, fmt.Errorf("invalid duration for %s=%q: %w", flagName, raw, err)
+	}
+	if d <= 0 {
+		return 0, fmt.Errorf("%s must be > 0 (got %q)", flagName, raw)
+	}
+	return d, nil
+}
+
 // ResourceConfig represents the structure of the configuration file.
 type ResourceConfig struct {
 	Resources []Resource `yaml:"resources"`
@@ -83,12 +100,35 @@ type ReleaseConfig struct {
 	Prefix   string            `yaml:"prefix"`
 	Labels   map[string]string `yaml:"labels"`
 	ChartRef ChartRefConfig    `yaml:"chartRef"`
-	// HelmInstallTimeout overrides the Flux HelmRelease Install.Timeout and
-	// Upgrade.Timeout for this Application kind. When zero, flux defaults
-	// apply. Populated from the
-	// release.cozystack.io/helm-install-timeout annotation on the
-	// ApplicationDefinition at start-up.
+	// HelmInstallTimeout is a per-Application override of Install.Timeout
+	// and Upgrade.Timeout. When non-zero, it wins over
+	// HelmReleaseInstallTimeout / HelmReleaseUpgradeTimeout below.
+	// Populated from the release.cozystack.io/helm-install-timeout
+	// annotation on the ApplicationDefinition at start-up.
 	HelmInstallTimeout time.Duration `yaml:"helmInstallTimeout,omitempty"`
+	// HelmReleaseInterval is the global default for Spec.Interval on
+	// HelmReleases generated by cozystack-api. Set from the api server
+	// --helmrelease-interval flag; mirrors the cozystack-operator flag of
+	// the same name so both HelmRelease-generating paths use the same
+	// reconcile cadence.
+	HelmReleaseInterval time.Duration `yaml:"helmReleaseInterval,omitempty"`
+	// HelmReleaseRetryInterval is the global default for
+	// Install/Upgrade.Strategy.RetryInterval. Decoupled from
+	// HelmReleaseInterval so failed install/upgrade retries recover fast
+	// without polling healthy releases at the same cadence.
+	HelmReleaseRetryInterval time.Duration `yaml:"helmReleaseRetryInterval,omitempty"`
+	// HelmReleaseInstallTimeout is the global default for
+	// Spec.Install.Timeout. Overridden per-Application by HelmInstallTimeout
+	// when the annotation is set.
+	HelmReleaseInstallTimeout time.Duration `yaml:"helmReleaseInstallTimeout,omitempty"`
+	// HelmReleaseUpgradeTimeout is the global default for
+	// Spec.Upgrade.Timeout. Overridden per-Application by HelmInstallTimeout
+	// when the annotation is set.
+	HelmReleaseUpgradeTimeout time.Duration `yaml:"helmReleaseUpgradeTimeout,omitempty"`
+	// HelmReleaseMaxHistory is the global default for Spec.MaxHistory.
+	// 0 means unlimited per Helm semantics; matches the cozystack-operator
+	// flag of the same name.
+	HelmReleaseMaxHistory int `yaml:"helmReleaseMaxHistory,omitempty"`
 }
 
 // ChartRefConfig references a Flux source artifact for the Helm chart.
