+++ pkg/config/config_test.go
@@ -117,3 +117,42 @@ func TestParseHelmInstallTimeoutAnnotation(t *testing.T) {
 		})
 	}
 }
+
+// ParsePositiveDuration is the startup validator shared by cozystack-operator
+// and cozystack-api. Both binaries reject zero/negative/malformed values so a
+// misconfigured flag fails fast instead of propagating into every generated
+// HelmRelease and being rejected by helm-controller's CRD validation later.
+func TestParsePositiveDuration(t *testing.T) {
+	tests := []struct {
+		name    string
+		raw     string
+		want    time.Duration
+		wantErr bool
+	}{
+		{name: "valid seconds", raw: "30s", want: 30 * time.Second},
+		{name: "valid minutes", raw: "5m", want: 5 * time.Minute},
+		{name: "valid compound", raw: "1h30m", want: 90 * time.Minute},
+		{name: "zero rejected", raw: "0s", wantErr: true},
+		{name: "negative rejected", raw: "-5m", wantErr: true},
+		{name: "malformed rejected", raw: "5x", wantErr: true},
+		{name: "empty rejected", raw: "", wantErr: true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ParsePositiveDuration("--test-flag", tt.raw)
+			if tt.wantErr {
+				if err == nil {
+					t.Fatalf("expected error for %q, got nil", tt.raw)
+				}
+				return
+			}
+			if err != nil {
+				t.Fatalf("unexpected error for %q: %v", tt.raw, err)
+			}
+			if got != tt.want {
+				t.Errorf("got %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
